requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
jinja2>=3.1.0

# Database
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Storage
aiofiles>=23.0.0
aioboto3>=12.0.0  # Optional: for S3/R2 storage backends

# Authentication (SaaS mode)
clerk-backend-api>=1.0.0
httpx>=0.24.0

# Billing (SaaS mode)
stripe>=7.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
ruff>=0.8.0
pre-commit>=4.0.0
